commands: move cli.go into commands/root.go
It'll facilitate testing cmdline
